+++ tests/python_tests/z_state/reconfig/test_math_reconfig.py
@@ -4,12 +4,12 @@
 from itertools import product
 
 from conftest import skip_for_blackhole, skip_for_wormhole
-from helpers.dump import TensixDump
 from helpers.format_config import DataFormat, FormatConfig
 from helpers.llk_params import (
     DestAccumulation,
 )
 from helpers.param_config import parametrize
+from helpers.tensix import TensixState
 from helpers.test_config import TestConfig
 from helpers.test_variant_parameters import CONFIGURE_TEST_RUN_IDX, TO_FROM_INT8
 
@@ -88,7 +88,8 @@ def test_math_reconfig(
         dest_acc=dest_acc,
     )
 
-    expected = configuration.run(workers_tensix_coordinates).dumps[0]
+    configuration.run(workers_tensix_coordinates)
+    expected = TensixState.fetch(workers_tensix_coordinates)
 
     configuration = TestConfig(
         "sources/state/reconfig/math_reconfig_test.cpp",
@@ -104,6 +105,7 @@ def test_math_reconfig(
         dest_acc=dest_acc,
     )
 
-    actual = configuration.run(workers_tensix_coordinates).dumps[0]
+    configuration.run(workers_tensix_coordinates)
+    actual = TensixState.fetch(workers_tensix_coordinates)
 
-    TensixDump.assert_equal(expected, actual)
+    TensixState.assert_equal(expected, actual)
